Insulin trapezoidal calculation is fixed for long-acting insulin
Long acting insulins with a >instantaneous time at the peak value, had an incorrect calculation, which is now fixed.
A calculation with a reversed sign is corrected. This was currently a no-op as it was corrected either immediately or later in the code, but without fixing, that would not be guaranteed to be the case forever.